Fixed issue with office folder being deleted out of temp after msi copies the files into the path

+++ Office-ProPlus-Deployment/Microsoft.ProPlus.InstallGenerator/InstallGenWPF/InstallOffice.cs
@@ -49,8 +49,12 @@ public class InstallOffice
         try
         {
             //remove directory here
-            if (Directory.Exists(@"C:\Windows\Temp\OfficeProPlus"))
-                Directory.Delete(@"C:\Windows\Temp\OfficeProPlus", true);
+            if (!Directory.GetCurrentDirectory().ToLower().Contains("~") &&
+                !Directory.GetCurrentDirectory().ToLower().Contains("tmp"))
+            {
+                if (Directory.Exists(@"C:\Windows\Temp\OfficeProPlus"))
+                    Directory.Delete(@"C:\Windows\Temp\OfficeProPlus", true);
+            }
             //MinimizeWindow();
 
             Initialize();
